Move the keyboard to the drivers it belongs with
It is an input device, not a viewer: it reads stdin and emits keystrokes,
and the package it sat in now holds only the dearpygui viewer.

Ticket: none package move on an open PR
